feat: restructure server to support multiple endpoints and add agents and sources endpoint
