+++ memgpt/web/server/routers/sources.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+from memgpt.connectors.storage import StorageConnector
+
+router = APIRouter()
+
+
+@router.get("/sources", tags=["sources"])
+async def available_sources():
+    return StorageConnector.list_loaded_data()
